made find latest added recipe method, and fixed all tests with it.
now new private method is introduced that finds maximum id of newly added Recipe, and this way we know for sure which recipe was added. I fixed accordingly all tests where this was used. And because i don't know whether max will be found, I assert in each test that maximum is not null